Update README introduction
As a project assignment from the Modern Developer academy, we were asked
to submit a pull request to a real repository as an exercise. In this
particular exercise, I have updated the introduction for the Tether
library to include a bit more information so that anyone could
understand the basic gist of this tool.