Check for presence of addEventListener
In some server side rendering environments window is present, presumably
for different modules to communicate via global variables. But
addEventListener is not present because there's no way for an event to
be fired.

In such environments tether causes rendering to fail because once it
checks for the presence of window, it assumes addEventListener is also
present.

This commit makes tether more cautious before calling addEventListener.