Fix positioning bug for disabled gpu optimizations
Issue observed when position of tether element is
set with the css properties top, left, bottom or
right.

Before writing the css to the element's
style their values are parsed to floats, presumably
related to the check against the current css value.
This incorrectly converts '5px' > 5, which is an
illegal css value and is ignored by the browser.

Conversion to float would seem to be incorrect and
is removed.